replace int with size_t for id everywhere

+++ infra/Variable.hpp
@@ -87,7 +87,7 @@ class Variable {
   std::string name_;
   std::vector<Field> fields_{};
   std::set<std::string> branch_names_{};
-  std::set<short> branch_ids_{};
+  std::set<size_t> branch_ids_{};
   mutable std::vector<double> vars_{}; //!
   std::function<double(std::vector<double>&)> lambda_{[](std::vector<double>& var) { return var.at(0); }};//!
